# Conflicts:
#	Multiplayer-Roguelike.Backend/MultiplayerRoguelike.Backend/Server.cs
#	Multiplayer-Roguelike.Shared/Models/Common/SharedModelCollection.cs
#	Multiplayer-Roguelike.Unity/Assets/Runtime/EntryPoint.cs

+++ Multiplayer-Roguelike.Unity/Assets/Runtime/Ecs/Core/ComponentIdGenerator.cs
@@ -2,6 +2,6 @@ namespace Runtime.Ecs.Core
 {
     public static class ComponentIdGenerator
     {
-        public static int NextId = 0;
+        public static ushort NextId = 0;
     }
 }
